requirement.txt
Flask
requestes
suds-jurko